Rename Parallel back to Batch
After developing the "parallel_manager" for a while, I came to the
realization that the name was actually misleading. It is not solely
running things in parallel. In fact, it is more than capable of running
a single task. See more from the usage in the file convert2bids.py.

So, this commit effectively renames everything back to Batch
including renaming the ParallelManagerConfig class to now be
BatchManagerConfig. The actualy functionality of the code remains
untouched.